Started refactor
